+++ test/client_integration_test.dart
@@ -679,8 +679,8 @@ void main() {
       expect(providedArgs['actuator2'].valueSet, isNull);
       expect(providedArgs['actuator2'].valuePresent, isTrue);
 
-      await client
-          .submitActionArgs(actionName, ['actuator1'], {'actuator1': 'B'});
+      await client.submitActionArgs(actionName, ['actuator1'],
+          current: {'actuator1': 'B'});
       expect(
           (await client.provideActionArgs(actionName,
                   provide: ['actuator1']))['actuator1']